Refactor AddTokenPage
