+++ client/src/pages/AddTokenPage/types.d.ts
@@ -1,0 +1,17 @@
+import { PublicAddressDoublet, FioNameItemProps } from '../../types';
+
+export type AddTokenProps = {
+  currentFioAddress: FioNameItemProps;
+  results: any;
+  loading: boolean;
+};
+
+export type FormValues = {
+  tokens: PublicAddressDoublet[];
+};
+
+export type ArrayErrorsProps = {
+  chainCode?: string;
+  tokenCode?: string;
+  publicAddress?: string;
+}[];
